deploy new app with dynamo DB fully working
