+++ .env.example
@@ -1,11 +1,18 @@
 ENV=dev
-DISCORD_CLIENT_SECRET=
+
 DISCORD_CLIENT_ID=
 DISCORD_BOT_TOKEN=
-FIREBASE_CONFIG={"projectId": "<Your Project ID>"}
-FIRESTORE_EMULATOR_HOST=localhost:8081
-USE_LOCAL_FIREBASE=true
-STARKNET_ID_CONTRACT_ADDRESS=0x033233531959c1da39c28daf337e25e2deadda80ce988290306ffabcd735ccbd
-STARKNET_ID_INDEXER_URL=https://indexer.starknet.id/fetch_tokens_id
-VERIFIER_DECIMAL_CONTRACT_ADDRESS=218957698842707292111176682338308570428481820353543328403027153649547919416
+
+STARKNET_ID_CONTRACT_ADDRESS=0x0798e884450c19e072d6620fefdbeb7387d0453d3fd51d95f5ace1f17633d88b
+STARKNET_ID_INDEXER_URL=https://goerli.indexer.starknet.id/field_data_to_id
+VERIFIER_DECIMAL_CONTRACT_ADDRESS=2858829565965467824506234522366406559425492229537050207406969294731822669741
 DISCORD_TYPE=28263441981469284
+
+AWS_PROFILE=
+AWS_REGION=
+DYNAMODB_TABLE_GUILD_DEV=
+DYNAMODB_TABLE_STARKNET_ID_DEV=
+
+JWT_KEY=
+RST_KEY=
+API_URL="http://localhost:3000/api"
