+++ apps/discord-bot/src/bot/commands/listRulesForRole.ts
@@ -1,5 +1,5 @@
 import { CommandInteraction, Client, SelectMenuInteraction } from 'discord.js';
-import { DiscordRule } from '../../dynamodb-libs/db-types';
+import { DiscordRule } from '../../dynamodb/db-types';
 import { getRulesForGuild } from '../../models/rule';
 import { formatRule, numberOfUserWithRole } from './utils';
