+++ apps/discord-bot/src/dynamodb/client.ts
