+++ apps/discord-bot/src/dynamodb/doc-client.ts
